Add syntax to do something useful
Just as a commented out entry. An alias can be defined for this, but
executes a git command when reloaded the aliases file, which is
something I don't want to happen.

So instead of defining an alias for this, I just keep a commented entry
documenting how to do it (just in case I need it again in the future).